requirements.txt
pandas
numpy
scikit-learn
yfinance
matplotlib
seaborn
pyyaml
pyarrow
pytest
